ROX-28473: Add tabs within VM Reporting
