+++ ui/apps/platform/src/Containers/Vulnerabilities/VulnerablityReporting/VulnReports/OnDemandReportsTab.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function OnDemandReportsTab() {
+    return <div />;
+}
+
+export default OnDemandReportsTab;
